% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Title}
\usage{
get_teams(includeRoster = TRUE)
}
\arguments{
\item{includeRoster}{flag to enable rosters for each team}
}
\value{
all teams from the NHL API
}
\description{
Title
}
\examples{
get_teams()

}
